==> 05-tidyverse2.Rmd <==
# Tidyverse 2 {#session5}

## Reflecting on the last session

We have managed to do quite a few tasks already. However, we also managed to create a proper mess in the process. Our **environment**  must be full of objects that we will not be using. And even though we could remove them it is best if we actually do not create them in the first place. That can be achieved with the use of the **pipe operator** or else `%>%` (**ctrl+shift+m**).

## Learning objectives   

1. Work more efficiently using tidyverse (dplyr and tidyr)
2. manipulating data
3. summarising data (in brief)


**Note:** In this session we will be focusing on describing a number of tasks and then discussing how we can best perform them. We have already introduced enough concepts to do really meaningful work. Now we have to put our progress to the test. 


## Pipe Operator `%>%`  

Lets start with seeing how the **pipe operator** works. If you think about the last exercise in the previous session you can remember that it took a number of steps where we kept assigning our work to a new data frame. With the use of `%>%` we can write all these steps one under the other in one chunk of code.

**Exercise 9**

Load the file **alcohol_exp** to a data frame called **df**. Then write the appropriate commands to complete the following steps:

1. Set group, session, meditation, and gender as factors
2. Keep only these four variables and response.rt
3. Recode session from 1 and 2 to **pre** and **post**
4. Recode group from 1 and 2 to **high** and **low**
5. Keep only Female participants
6. Keep only response.rt below 1.5
6. save your final data frame into a file called **"wastedresources.csv"**


```{r eval=FALSE}
library(tidyverse)
df <- read_csv("alcohol_exp.csv")

df3 <- df %>% 
  mutate(group = as.factor(group)) %>% 
  mutate(session = as.factor(session)) %>% 
  mutate(meditation = as.factor(meditation)) %>% 
  mutate(gender = as.factor(gender)) %>% 
  select(group, session, meditation, gender, response.rt) %>% 
  filter(gender == 2) %>% 
  filter(response.rt < 1.5) %>% ### this could be done with the above
  mutate(gender = recode(gender,'1' = "Male", '2'="Female")) %>% 
  mutate(group = recode(group,'1' = "High", '2'="Low"))



### and just for fun lets also create a histogram of reaction times.
hist(df3$response.rt)

### lets save the data file

write_csv(df3, "wastedresources.csv")
  
```

## An example of a FYP data analysis   

Let's assume that the datafile **alcohol_qualtrics.csv** (you can find it on Moodle in the fRom zeRo to heRo section) has been generated by Qualtrics and we want to go through the analysis with our FYP students. The process of going through the analysis will vary for different projects but in general we can assume the three main steps are the following:   

1. Cleaning the data   
2. Creating new variables (e.g., overall scores for questionnaires)   
3. Carrying out the main analysis   

## A brief overview of our data  

1. **id** is the participant code   
2. **participant** is also our participant code   
3. **gender** 1: male, 2:female   
4. **Age** in years   
5. **Days1** to **Days4** are items of the **Days** questionnaire that is scored as the sum of its items (Likert 1-5)   
6. **Smoke** whether a participant smokes tobacco, 1: Yes, 2:No   
7. **Drinks** the number of days of the week that the participant consumes alcohol   
8. **Years** the number of years that the participant has been consuming alcohol   
9. **Freq1** to **Freq4** items of the **Frequency** questionnaire that is scored as the mean of its items (Likert 1-5)   
10. **Units** number of alcohol units consumed in a week  

The rest of the columns are irrelevant and we will not be working with them. Before we carry out any analysis we want to prepare our data file. Follow the steps described in the exercise below to prepare your datafile.

## Exercise 6.1

1. Load the file **"alcohol_qualtrics.csv"** to a data frame
2. Drop all the variables after the column 17 (drop column 17 as well)
3. Remove any participants with participant number above 100 and below 300   
4. Make sure gender and smoke are factors
5. Add a column that is the sum of Days1 to Days4
6. Reverse code the variable Freq4
7. Add a column that is the mean of Freq1, Freq2, Freq3, and the above reversed coded variable
8. Save the final file using a file name of your choice.

```{r}
library(tidyverse)
library(rstatix)


df <- read_csv("alcohol_qualtrics.csv")


df2 <- df %>% 
  select(1:17) %>% 
  filter(participant>100 & participant<300) %>% 
  mutate(Days = Days1 + Days2 + Days3 + Days4) %>% 
  mutate(Freq4_r = recode(Freq4, '1'=5, '2'=4, '3'=3, '4'=2, '5'=1)) %>% 
  mutate(Frequency = (Freq1+Freq2+Freq3+Freq4_r)/4) %>% 
  mutate(gender=as.factor(gender)) %>% 
  mutate(Smoke=as.factor(Smoke))

write_csv(df2, "exercise61.csv")
```



## Working through our participants section   
Typically in our participants section we want to provide some descriptive figures for age and frequencies per gender. 


```{r demographics, message=FALSE, warning=FALSE}
library(tidyverse)
library(rstatix)  

df <- read_csv("exercise61.csv")

df$gender <- as.factor(df$gender)


### Overall demographics  

df %>% 
  get_summary_stats(Age)

df %>% 
  count(gender)

```


In some cases we might also want to further break down that information  per group. Below we can see how we can achieve that.  


```{r, message=FALSE, warning=FALSE}
### If you havent done so make sure Smoke is a factor

df %>% 
  group_by(Smoke) %>% 
  get_summary_stats(Age)

df %>% 
  group_by(Smoke) %>% 
  count(gender)


```



## Exercise 6.2  

Acquire descriptive figures (overall and per Smoke group) for the variables **Drinks**, **Years**, **Frequency**, **Days**, and **Units**.  


## Running a simple correlational analysis for Frequency and Days

We can now carry out a simple correlational analysis. This will be just as means of demonstration. There will be a dedicated session on Correlation and Regression. As we have already prepared our Frequency and Days variables we can proceed with the correlation function.   

```{r, eval=FALSE}
###  cor_test is part of the rstatix package

df %>% 
  cor_test(Frequency, Days, method = "pearson")

```


The **cor_test()** function offers a variety of arguments that will allow us to run all sorts of correlation combination. Let us see the **help** for this function. We can access the help is built-in within R in a number of different ways. We could place our cursor in the function we need help for and press F1. Or at the bottom right pane we can click on Help and then type the name of the function. 


## In conclusion  

Carrying out a FYP data analysis can be broken down to smaller pieces and generally this approach can be quite helpful. Always try to have different scripts for data preparation, demographics, and your main data analysis. This will also make it easier to recycle through your scripts and use them across different projects.
