==> 08-Restructuring-and-Merging.Rmd <==
# Restructuring and Merging Data. {#session9}  


## Learning Objectives  
In this session we will learn how to change from **wide** to **long** data and vice versa. We will also learn how to merge two different datasets and carry out a correlation analysis.  

## Restructuring Data

As we have seen so far we have been working with datasets in **long** format. Meaning we have different columns to denote our variables and our dependent variable only in one column. However, there will be cases where we will need to use what is called a **wide** format where we will need to have measures across different columns.  
For example, in the previous stroop paradigm we may want to have one column for the **salient** reaction times and one column for the **neutral* reaction times. One reason why we may need this is because we may want to calculate a new variable that will be the difference between the two variables. That way we will end up having one row per participant which will make our dataset ideal for merging with another dataset that contains questionnaires responses. Our goal for this session will be to merge experimental and survey data in one file and carry out correlational analysis.

For restructuring our data we will be using **pivot_wider()** and **pivot_longer()**. Both are part of tidyverse and specifically **tidyr**. They are new implementations and will probably replace **spread()** and **gather()**. Most tutorials on the web use spread and gather. We will use both the new and old approach.

Let us continue our work from the previous session and work with the file "stroopfinal".

```{r eval=FALSE}
library(tidyverse)

dflong <- read_csv("stroopfinal.csv")

```


We will now convert our dataframe from long to wide format. We will do using two different ways. Spread() and gather() are being depreciated but we still use them as an example.

```{r eval=FALSE}

## we now want to bring the file to a form that will have two RTs columns
## one for salience trials and one for neutral trials
dfwide <- dflong %>% 
  pivot_wider(names_from = salience, values_from = RT)

## we could do the same with spread()
dfwide2 <- dflong %>% 
  spread(salience, RT)

```


Just for practice we can revert back from wide to long format.  

```{r eval=FALSE}

## Let us now try to convert the wide file back to long

dflong2 <- dfwide %>% 
  pivot_longer(cols = c(Gambling, Neutral), 
               names_to = "Salience", 
               values_to = "Reaction")
### It would be very similar using gather

dflong3 <- dfwide %>% 
  gather(key = "Salience", value = "Reaction", Gambling, Neutral)

## dflong2 and dflong3 look different as they are sorted by different variable
## this is just a visual difference but we can fix it anyway

dflong3 <- dflong3 %>% 
  arrange(participant, group)
```


As I said we do not need to revert back to the long format. We will carry on working with the wide format and calculate our **diff** variable.  
```{r eval=FALSE}
dfwide <- dfwide %>% 
  mutate(diff = Gambling - Neutral)

```


It would be very interesting now to see whether this difference correlates with the frequency of games. Let's also load the questionnaire.csv and merge it with our widedf.  

```{r eval=FALSE}
## load our datafile
questionnaire <- read_csv("questionnaire.csv")

## let's merge the two datasets dfwide and questionnaire
## first lets do it the wrong way

dfall <-left_join(dfwide, questionnaire)

## this will generate an error as group in questionnaire is not a factor
## also group in the questionnaire is 1 and 2 not the same as in dfwide
## also left_join() is not the appropriate function for our case
questionnaire <- questionnaire %>% 
  mutate(group = recode(group,'1' = "gambling", '2'="control"),
         group = as.factor(group))

dfwide$group = as.factor(dfwide$group)


## let's merge the two datasets dfwide and questionnaire

dfall <-full_join(dfwide, questionnaire)
```

That worked fine, even though we have some missing values. Let's see now if we have any significant correlation between the stroop difference and the frequency of games.  

Let' s start with a simple scatterplot.

```{r eval=FALSE}
scatter <- ggplot(dfall, aes(x = games, y = diff)) +
  geom_point() 

scatter
```


This looks rather unexciting, perhaps group can make things more interesting. 

```{r eval=FALSE}
scatter <- ggplot(dfall, aes(x = games, y = diff, color = group)) +
  geom_point() 

scatter

```

Let's carry out the correlational analysis next. 

```{r eval=FALSE}
library(Hmisc)
dfgambling <- dfall %>% 
  filter(group == "gambling")
rcorr(dfgambling$games, dfgambling$diff)

```
